backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sentence-transformers>=2.7.0
langdetect>=1.0.9
firebase-admin>=6.5.0
python-levenshtein>=0.25.0
scikit-learn>=1.4.0
numpy>=1.26.0
pydantic>=2.7.0
